Extract PuTTYKeyPairParser
Same treatment as Pkcs8KeyPairParser: the PuTTY .ppk parsing (header
parsing, passphrase decryption, MAC verification, key-pair construction)
moves into a PuTTYKeyPairParser behind the KeyPairParser strategy.
PuTTYKeyFile drops from ~380 lines to ~45 - readKeyPair() delegates, and
getType() / isEncrypted() / getKeyFileVersion() forward to the parser.

Unlike Pkcs8KeyPairParser the PuTTY parser is stateful: it retains the
parsed headers so the PuTTY-specific accessors can answer. That state is
"the parsed file" and the parser is used by a single provider instance,
so this matches how PuTTYKeyFile already worked.

No behaviour change: the parsing code moved verbatim.
